[frontend] Basic Diff Viewer implemented
* Difference viewer call was handled into button processing
* Application was divided into three modules
* Docker file minor changes
