+++ womparator/frontend/Dockerfile
@@ -6,7 +6,7 @@ COPY ./package.json .
 COPY ./package-lock.json .
 
 # install application dependencies
-RUN npm install
+RUN npm install --legacy-peer-deps
 
 # copy app files
 COPY . .
