+++ womparator/frontend/package.json
@@ -13,7 +13,9 @@
     "@testing-library/user-event": "^13.5.0",
     "emotion": "^11.0.0",
     "react": "^18.2.0",
+    "react-diff-viewer": "^3.1.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.8",
     "web-vitals": "^2.1.4"
